+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+* [Example] Simplify and update the example.
+
 ## [0.0.8] - 2020-06-14
 
 ### Added
@@ -198,6 +204,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 * Extraction from my personal configuration framework.
 
+[Unreleased]: https://github.com/ejpcmac/confkit/compare/master...develop
 [0.0.8]: https://github.com/ejpcmac/confkit/compare/v0.0.7...v0.0.8
 [0.0.7]: https://github.com/ejpcmac/confkit/compare/v0.0.6...v0.0.7
 [0.0.6]: https://github.com/ejpcmac/confkit/compare/v0.0.5...v0.0.6